move file and fix triggers
