+++ simgui.json
@@ -198,7 +198,7 @@
             0.0,
             0.8500000238418579
           ],
-          "height": 140,
+          "height": 354,
           "series": [
             {
               "color": [
